Added dateformat decorator.
The dateformat decorator was introduced so string date values can be
parsed to properties of datetime type.
Also included a custom exception that will be raised in cases where
the dataclass have been defined with a property of type datetime and
it hasn't been decorated with the dateformat decorator specifying the
dateformat to be used when parsing the string dates.