Support for global and field metadata
Previouly, there was a decorator dateformat to configure how the
datetime strings would be passed into datetime objects to be assigned
to properties in the class. However, there are times that the data in
a CSV file can have different formats. This commit addresses this issue
and read dateformat value from the metadata property of the Field
object. This will allow specifying how the datetime string value
should be parsed specifically for a property.

In the same way it has been also added a new decorator called
allow_whitespaces, which will allow values containing only white
spaces being assigned to the class string properties. This option can
also be set on a field level using the metadata property.

Last change as the removal of the custom exception classes that was
not longer used.