Small fix when parsing boolean values
When converting a value to boolean we use the strip function to remove
possible spaces that might be in the way of the data
conversion. However, when setting the value to 1 an error is raised
since the strip function is not available for integers.
This is usually not a problem when the value comes from a column in
the CSV file, but it is a problem when the value is set as the default
value of a boolean dataclass field.