Raise an error when mapped column is not found
Improve the error message when a mapped column is noit found in the CSV
file. A new exception type 'MappedColumnNotFoundError' has been introduced for
these cases.

Added additional tests to reproduce two different scenarios, the first
when a mapped column is not found and the second and a non-mapped column
is not found.